Make git a little quieter
On my system whenever I run the tests without specifying an initial branch name,
I get a lot of messages about setting the default branch name globally. This is
just noise for these tests, so this specifies a branch name manually to make it
go away.

The name ‘gac-test’ was chosen to make it clear that we're dealing with a
repository created specifically for ‘git-auto-commit-mode’ in case somehow a
repository gets left behind.